requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0

# Transformers and PEFT
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0

# Numerical/Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# Optional: Logging
wandb>=0.15.0
tensorboard>=2.14.0

# Optional: Quantization (for larger models)
# bitsandbytes>=0.41.0  # Uncomment if using quantization
